+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The types of changes are:
 ### Added
 - Adds support for custom get experiences fn and custom patch notices served fn [#4410](https://github.com/ethyca/fides/pull/4410)
 
+### Fixed
+- User preferences from cookie should always override experience preferences [#4405](https://github.com/ethyca/fides/pull/4405)
+
 ## [2.24.0](https://github.com/ethyca/fides/compare/2.23.3...2.24.0)
 
 ### Added